Tue, Jan 30, 2024, 2:02 PM -05:00

+++ bitcoin.py
@@ -1,4 +1,4 @@
 
-number = int(input("Enter number of Bitcoin: "))
+number = float(input("Enter number of Bitcoin: "))
 amount = number * 38761.0833
 print(float(amount))
